Spin more aggressively
Rather than trying to go to sleep immediately when we fail to acquire
the lock (and spinning only when we also fail to acquire the spinlock),
spin a few times before trying to go to sleep when there are no other
waiters.  This can improve performance when we only have a small number
of threads.